Chatbot/requirements.txt
os
json
datetime
nltk
ssl
streamlit
pillow
random
scikit-learn
pandas
gTTS